Enable VMR sync from private GitHub repo and AzDO repo
