+++ src/Microsoft.DotNet.Darc/tests/Microsoft.DotNet.DarcLib.Tests/VirtualMonoRepo/RepositoryCloneManagerTests.cs
@@ -25,9 +25,9 @@ public class RepositoryCloneManagerTests
     private readonly Mock<ILocalGitRepo> _localGitRepo = new();
     private readonly Mock<IProcessManager> _processManager = new();
     private readonly Mock<IFileSystem> _fileSystem = new();
-    private readonly Mock<IRemoteFactory> _remoteFactory = new();
-    private readonly Mock<IRemote> _remote = new();
-
+    private readonly Mock<IClonableRepoFactory> _remoteFactory = new();
+    private readonly Mock<IClonableGitRepo> _remote = new();
+    
     private RepositoryCloneManager _manager = null!;
 
     [SetUp]
@@ -53,7 +53,7 @@ public class RepositoryCloneManagerTests
 
         _remote.Reset();
         _remoteFactory.Reset();
-        _remoteFactory.SetReturnsDefault(Task.FromResult(_remote.Object));
+        _remoteFactory.SetReturnsDefault(_remote.Object);
 
         _manager = new RepositoryCloneManager(
             _vmrInfo.Object,
@@ -74,7 +74,7 @@ public class RepositoryCloneManagerTests
         path = await _manager.PrepareClone(RepoUri, "main", default);
         path.Should().Be(ClonePath);
 
-        _remoteFactory.Verify(x => x.GetRemoteAsync(RepoUri, It.IsAny<ILogger>()), Times.Once);
+        _remoteFactory.Verify(x => x.GetRemote(RepoUri, It.IsAny<ILogger>()), Times.Once);
         _remote.Verify(x => x.Clone(RepoUri, Ref, ClonePath, false, null), Times.Once);
         _localGitRepo.Verify(x => x.Checkout(ClonePath, "main", false), Times.Exactly(2));
     }
@@ -91,7 +91,7 @@ public class RepositoryCloneManagerTests
         path = await _manager.PrepareClone(RepoUri, "main", default);
         path.Should().Be(ClonePath);
 
-        _remoteFactory.Verify(x => x.GetRemoteAsync(RepoUri, It.IsAny<ILogger>()), Times.Never);
+        _remoteFactory.Verify(x => x.GetRemote(RepoUri, It.IsAny<ILogger>()), Times.Never);
         _remote.Verify(x => x.Clone(RepoUri, Ref, ClonePath, false, null), Times.Never);
         _processManager.Verify(x => x.ExecuteGit(ClonePath, "fetch", "--all"), Times.Once);
         _localGitRepo.Verify(x => x.Checkout(ClonePath, Ref, false), Times.Once);
